Enable the library to work with any TwoWire and not just Wire
This update makes a backward compatible version that allows the use of any TwoWire
This is especially useful on an Arduino UNO R4 WiFi that has a second Stemma QT connected I2C bus on 3.3V, named "Wire1", that works well with most Wii controller adapters. In my case it is tested on the original Nintendo Nunchuck and the Nyko Kama wireless nunchuck.
This update also introduces a new function named isConnected() that can be used to detect is the device is responding to I2C traffic (e.g when the wireless link is up)
1 parent